+++ examples/ExampleBrowser/CMakeLists.txt
@@ -175,6 +175,8 @@ SET(BulletExampleBrowser_SRCS
 	../SharedMemory/b3PluginManager.cpp
 	../RobotSimulator/b3RobotSimulatorClientAPI.cpp
 	../RobotSimulator/b3RobotSimulatorClientAPI.h
+	../RobotSimulator/b3RobotSimulatorClientAPI_NoGUI.cpp
+	../RobotSimulator/b3RobotSimulatorClientAPI_NoGUI.h
 	../BasicDemo/BasicExample.cpp
 	../BasicDemo/BasicExample.h
 	../InverseDynamics/InverseDynamicsExample.cpp
